perf/core: Implement workaround to add pmu stop before unthrottling to prevent WARNING
It is a temporary workaround for ast530.
We will continue to investigate and resolve the root cause in the future.

Currently, during the perf sampling, if the perf interrupt takes too
long, perf framework will lower the perf_event_max_sample_rate.
This will limit the number of samples per timer tick (max_samples_per_tick)
and set hwc->interrupts to MAX_INTERRUPTS
within the __perf_event_account_interrupt() function.

Afterward, the perf framework will unthrottle the event in the timer
interrupt handler, which triggers the driver's *_pmu_start() function.
Most of the driver's *_pmu_start() functions will check the event->hw.state
to determine whether this event has stopped. If the event has not stopped,
a WARN_ON_ONCE() warning will be triggered.

This is currently a workaround and may be reverted in the future.

Reformed from the following:
https://patchwork.kernel.org/project/linux-riscv/patch/20230602094841.1225-1-eric.lin@sifive.com/

Signed-off-by: Locus Wei-Han Chen <locus84@andestech.com>